requirements.txt
asgiref==3.7.2
astunparse==1.6.3
black==23.12.1
click==8.1.7
defusedxml==0.7.1
Django==4.2.9
django-click==2.3.0
django-extensions==3.2.3
djangorestframework==3.14.0
drf-writable-nested==0.7.0
et-xmlfile==1.1.0
html-json-forms==1.1.1
importlib-metadata==7.0.1
mypy-extensions==1.0.0
natural-keys==2.1.0
openpyxl==3.0.9
packaging==23.2
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.1.0
psycopg2-binary==2.9.9
pystache==0.6.5
pytz==2023.3.post1
pyxform==2.0.0
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
wq==2.0.0
wq.app==2.0.0
wq.build==2.0.0
wq.create==2.0.0
wq.db==2.0.0
xlrd==2.0.1
xlsconv==2.0.0
zipp==3.17.0
drf_yasg==1.21.7
mysqlclient==2.2.1